Timers working

+++ protocols/rscp/mtime.c
@@ -3,7 +3,7 @@
 
 void mtime_get_current(mtime *t) {
  t->seconds = clock_get_time();
- t->millis = clock_get_ticks() * MTIME_MILLIS_PER_TICK;
+ t->millis = (uint16_t) clock_get_ticks() * MTIME_MILLIS_PER_TICK;
 }
 
 int8_t mtime_compare(const mtime *a, const mtime *b) {
